# plugwerk-api/src/main/resources/openapi/plugwerk-api.yaml
openapi: 3.1.0
info:
  title: Plugwerk API
  description: |
    **Plugwerk** is a self-hosted plugin marketplace for the [PF4J](https://pf4j.org/) ecosystem.
    It lets teams publish, version, and distribute Java/Kotlin plugins to their own applications
    without relying on a public registry.

    ## Core Concepts

    ### Namespaces
    A **namespace** is the top-level organisational unit. Every plugin belongs to exactly one
    namespace. Namespaces are identified by a URL-safe **slug** (lowercase alphanumeric + hyphens,
    2–64 characters). You might use one namespace per product, team, or customer, e.g. `acme-core`.

    ### Plugins
    A **plugin** is a logical grouping of releases for a single PF4J plugin ID. The `pluginId`
    matches the `Plugin-Id` entry in the PF4J manifest (`MANIFEST.MF` or `plugin.properties`).
    Each plugin can have a human-readable name, description, icon, and categorisation metadata.

    ### Releases
    A **release** is a specific versioned artifact (JAR or ZIP) for a plugin. Versions follow
    [Semantic Versioning](https://semver.org/) (`MAJOR.MINOR.PATCH`). Releases go through a
    lifecycle: `draft` → `published` → `deprecated` / `yanked`. Only `published` releases are
    returned to catalog and update-check consumers. A namespace owner can optionally require
    manual review before a release is published (see the **Reviews** tag).

    ### Plugin Descriptor (MANIFEST.MF)
...

    | Attribute | Required | Description |
...

...

...

...

...
    ```
...

...

...
servers:
...
security: []
...
paths:
...
  /admin/storage/consistency:
    get:
      tags:
        - AdminStorageConsistency
      summary: Scan storage vs DB and return inconsistencies (#190)
      description: |
        Walks the entire `plugin_release` table and the configured storage
        backend, returning a [`ConsistencyReport`](#/components/schemas/ConsistencyReport)
        with two lists:

        * `missingArtifacts`: DB rows whose `artifactKey` does not exist in
          storage (caused by storage outage during upload, accidental
          delete in the storage console, or an incomplete `fs ↔ s3`
          migration).
        * `orphanedArtifacts`: storage objects with no DB row pointing at
          them (caused by best-effort post-commit storage cleanup
          failing — see #481).

        Synchronous and uncached. A `max-keys-per-scan` circuit breaker
        (default 100 000, see
        `plugwerk.storage.consistency.max-keys-per-scan`) aborts the scan
        with `409 CONFLICT` rather than hanging the admin UI on a runaway
        bucket. An async-scan endpoint is **out of scope** for the first
        iteration.

        Requires superadmin.
      operationId: getStorageConsistencyReport
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Consistency snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsistencyReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Scan aborted by the max-keys-per-scan circuit breaker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/storage/consistency/releases/{releaseId}:
    delete:
      tags:
        - AdminStorageConsistency
      summary: Remove a `plugin_release` row whose artifact is missing (#190)
      description: |
        Deletes the DB row and best-effort the storage object (idempotent —
        re-running on a vanished row succeeds with 204). Used to clear
        `missingArtifacts` entries from the consistency report. Requires
        superadmin.
      operationId: deleteOrphanedRelease
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: releaseId
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the orphaned release row to remove.
      responses:
        '204':
          description: Row removed (or already gone).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/storage/consistency/artifacts:
    delete:
      tags:
        - AdminStorageConsistency
      summary: Bulk-delete orphaned storage objects (#190)
      description: |
        Deletes the supplied storage keys after re-checking that no
        `plugin_release` row references them inside the same transaction.
        Returns the partition of deleted vs skipped keys (skipped =
        became referenced between scan and delete, or storage backend
        failed).

        Bypasses the reaper's grace-period guard because the admin is
        making an explicit choice to delete. The narrow TOCTOU window
        between the recheck and the storage `delete` is accepted as the
        admin-action cost. Requires superadmin.
      operationId: deleteOrphanedArtifacts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrphanedArtifactDeletionRequest'
      responses:
        '200':
          description: Partition of deleted vs skipped keys.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkArtifactDeletionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
...
  schemas:
...
    ConsistencyReport:
      type: object
      description: |
        Bidirectional reconciliation snapshot between `plugin_release` rows
        and the configured storage backend (#190).
      properties:
        missingArtifacts:
          type: array
          items:
            $ref: '#/components/schemas/MissingArtifact'
        orphanedArtifacts:
          type: array
          items:
            $ref: '#/components/schemas/OrphanedArtifact'
        scannedAt:
          type: string
          format: date-time
          description: Wall-clock timestamp the scan started.
        totalDbRows:
          type: integer
          format: int32
          description: Distinct artifact keys across `plugin_release`.
        totalStorageObjects:
          type: integer
          format: int32
          description: Objects walked in storage (subject to the max-keys circuit breaker).
      required:
        - missingArtifacts
        - orphanedArtifacts
        - scannedAt
        - totalDbRows
        - totalStorageObjects
    MissingArtifact:
      type: object
      description: DB row whose `artifactKey` does not exist in storage.
      properties:
        releaseId:
          type: string
          format: uuid
        pluginId:
          type: string
        version:
          type: string
        artifactKey:
          type: string
      required:
        - releaseId
        - pluginId
        - version
        - artifactKey
    OrphanedArtifact:
      type: object
      description: |
        Storage object with no `plugin_release` row pointing at it.
        `ageHours` is computed at scan time against `lastModified` so the
        UI renders deterministic values across re-renders; the reaper
        (#496) uses the same field to enforce its grace-period guard.
      properties:
        key:
          type: string
        lastModified:
          type: string
          format: date-time
        ageHours:
          type: integer
          format: int64
          description: Whole hours since the object was last modified at scan time.
        sizeBytes:
          type: integer
          format: int64
      required:
        - key
        - lastModified
        - ageHours
        - sizeBytes
    OrphanedArtifactDeletionRequest:
      type: object
      description: Bulk-delete payload for orphaned storage keys.
      properties:
        keys:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            type: string
      required:
        - keys
    BulkArtifactDeletionResult:
      type: object
      description: |
        Partition of the requested keys into successfully deleted vs
        skipped (either re-referenced between scan and delete, or storage
        backend failed).
      properties:
        deleted:
          type: array
          items:
            type: string
        skipped:
          type: array
          items:
            type: string
      required:
        - deleted
        - skipped
    ErrorResponse:
      type: object
      description: |
        Standard error response body returned for all 4xx and 5xx
        responses. The `status`, `error`, and `message` fields vary by
        endpoint and failure mode — schema-level examples would be
        misleading because the same envelope is reused for everything
        from `400 BAD_REQUEST` validation failures to `503 SERVICE_UNAVAILABLE`
        infrastructure outages. Per-response `examples` blocks on
        individual endpoints carry the realistic shape for that case.
      properties:
        status:
          type: integer
          description: HTTP status code (mirrors the response status).
        error:
          type: string
          description: |
            Short machine-readable error code (e.g. `NOT_FOUND`,
            `BAD_REQUEST`, `SERVICE_UNAVAILABLE`).
        message:
          type: string
          description: Human-readable description of what went wrong.
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp of when the error occurred.
      required:
        - status
        - error
        - message
        - timestamp
...
  responses:
...
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT token obtained via `POST /api/v1/auth/login`.
        Include as `Authorization: Bearer <token>`. Default validity: 8 hours.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Long-lived API key for CI/CD pipelines and service accounts.
        Include as `X-Api-Key: <your-api-key>` header.
        Namespace-scoped access keys are provisioned by a namespace ADMIN;
        each key grants implicit ADMIN access to its issuing namespace.